trying my best to decode.  Been sleepless for 3 nights

+++ where_collatz_is.h
@@ -3,5 +3,5 @@
 
 
 unsigned long long int randomizingNum(int max, int min);
-int collatzingUntilOne(unsigned long long int randomNum);
+unsigned long long int collatzingUntilOne(unsigned long long int randomNum);
 #endif
